requirements.txt
# Kryptopedia requirements file
fastapi==0.104.1
uvicorn==0.23.2
motor==3.3.1
pydantic==2.4.2
pydantic[email]==2.4.2
python-multipart==0.0.6
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
jinja2==3.1.2
python-dateutil==2.8.2

# Optional dependencies - uncomment to enable as needed
# boto3==1.28.65  # For S3 storage
# elasticsearch==8.10.0  # For Elasticsearch
# redis==5.0.1  # For Redis caching
